Fixed Health Bars
PlayerHealth.cs was being run before the health bar had a chance to initialize.

So the PlayerHealth script would try to call SetMaxHealth() on the HealthBar.cs script, but the HealthBar.cs script was unable to set the maximum health since the slider hasn't been initalized yet. So here's how I fixed it.

- Changed script execution order in project settings so that PlayerHealth.cs runs before HealthBar.cs

Extra changes:
- Added health foreground and background sprite to health bars that were missing them.